Small fixes and renaming file
