+++ src/application/services/shortener.service.ts
